Converted key parameters from integers to doubles
- Converted ball positions to doubles
- Experimented with functions to avoid clipping
- Success for lines and points, more work required for balls

+++ pong-clone/inc/field.hpp
@@ -31,11 +31,11 @@ struct BallArgs
 {
    int radius;
 
-   int init_x;
+   double init_x;
 
-   int init_y;
+   double init_y;
 
-   float speed;
+   double speed;
 };
